src/ai_services/clip/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
fastapi>=0.100.0
python-multipart>=0.0.6
uvicorn>=0.22.0
pillow>=9.5.0
numpy>=1.24.0
pydantic>=2.0.0 